HOSTDEV-1466 another fix for supporting our default.in by as default input
This commit fixes the following issue:
Although the version.py file is now generated correctly without the input in buildout.cfg now, in some projects the development-scripts step fails.

I'm not sure what is trying to reference the section or why it fails, but this is the error:
{noformat}
While:
  Installing development-scripts.
  Getting option __version__.py:inline.
  Getting section infi.recipe.template.version.
Error: The referenced section, 'infi.recipe.template.version', was not defined.

{noformat}